fix(john): inconsistent paddings in chatarea
