+++ coffeeAGNTCY/coffee_agents/lungo/frontend/src/components/Chat/Message.tsx
@@ -3,48 +3,31 @@
  * SPDX-License-Identifier: Apache-2.0
  **/
 
-import React, { useEffect, useRef, useState } from "react"
+import React from "react"
 import { Box, Stack, Typography } from "@open-ui-kit/core"
-import Person from "@mui/icons-material/Person"
-import SmartToy from "@mui/icons-material/SmartToy"
-import { LoadingDots } from "@/components/loading"
+import type { SystemStyleObject } from "@mui/system"
+import type { Theme } from "@mui/material/styles"
 
-interface SlowTextProps {
-  text: string
-  speed?: number
+const messageBodyTypographySx: SystemStyleObject<Theme> = {
+  whiteSpace: "pre-wrap",
+  overflowWrap: "break-word",
+  wordBreak: "break-word",
+  py: 1,
 }
 
-const SlowText: React.FC<SlowTextProps> = ({ text, speed = 25 }) => {
-  const [displayedText, setDisplayedText] = useState<string>("")
-  const idx = useRef<number>(-1)
-
-  useEffect(() => {
-    function tick(): void {
-      idx.current++
-      setDisplayedText((prev: string) => prev + text[idx.current])
-    }
-
-    if (idx.current < text.length - 1) {
-      const addChar = setInterval(tick, speed)
-      return () => clearInterval(addChar)
-    }
-  }, [displayedText, speed, text])
-
-  return <span>{displayedText}</span>
-}
-
-interface MessageProps {
-  content: string
-  aiMessage: boolean
-  animate: boolean
-  loading: boolean
+export interface MessageProps {
+  /** Leading avatar or icon for the row. */
+  icon: React.ReactNode
+  /** Message body; wrapped in shared `body2` typography styles. */
+  children: React.ReactNode
+  /** Assistant / highlighted rows use `action.hover` background. */
+  highlighted?: boolean
 }
 
 const Message: React.FC<MessageProps> = ({
-  content,
-  aiMessage,
-  animate,
-  loading,
+  icon,
+  children,
+  highlighted = false,
 }) => {
   return (
     <Stack
@@ -53,22 +36,18 @@ const Message: React.FC<MessageProps> = ({
       spacing={2}
       sx={{
         width: "100%",
-        px: { xs: 2, sm: 4, md: 8, lg: 15 },
-        py: { xs: 3, md: 3.75 },
-        bgcolor: aiMessage ? "action.hover" : "transparent",
+        bgcolor: highlighted ? "action.hover" : "transparent",
       }}
     >
       <Box
         sx={{
-          width: 35,
-          height: 35,
           flexShrink: 0,
           display: "flex",
           alignItems: "center",
           justifyContent: "center",
         }}
       >
-        {aiMessage ? <SmartToy /> : <Person />}
+        {icon}
       </Box>
       <Box
         sx={{
@@ -78,21 +57,13 @@ const Message: React.FC<MessageProps> = ({
           wordBreak: "break-word",
         }}
       >
-        {loading ? (
-          <LoadingDots size={12} />
-        ) : (
-          <Typography
-            variant="body2"
-            component="div"
-            sx={{
-              whiteSpace: "pre-wrap",
-              overflowWrap: "break-word",
-              wordBreak: "break-word",
-            }}
-          >
-            {animate ? <SlowText speed={20} text={content} /> : content}
-          </Typography>
-        )}
+        <Typography
+          variant="body2"
+          component="div"
+          sx={messageBodyTypographySx}
+        >
+          {children}
+        </Typography>
       </Box>
     </Stack>
   )
